requirements.txt
discord.py
python-dotenv
requests
youtube_dl